src/requirements.txt
python-libnmap==0.7.3
redis==4.6.0
requests==2.31.0
rq==1.16.2
tailer==0.4.1
elasticsearch==7.17.9
manuf==1.1.5
pysafebrowsing==0.1.4
